requirements.txt
pdfplumber
pydantic
regex
groq
python-dotenv
